# apps/web/public/admin/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SMM Enduro Challenge API",
    "version": "0.1.0",
    "description": "API contract for the Santa Monica Mountains Enduro Challenge serverless backend."
  },
  "servers": [
    {
      "url": "https://d152gxg9dh92dl.cloudfront.net/api",
      "description": "CloudFront API entrypoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "Challenge": {
        "type": "object",
        "required": ["id", "name", "description", "startDate", "endDate", "status", "segmentIds", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["DRAFT", "ACTIVE", "COMPLETED"] },
          "segmentIds": { "type": "array", "items": { "type": "string", "format": "uuid" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Segment": {
        "type": "object",
        "required": ["id", "stravaSegmentId", "name", "challengeId", "distance", "elevationGain", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "stravaSegmentId": { "type": "integer" },
          "name": { "type": "string" },
          "challengeId": { "type": "string", "format": "uuid" },
          "distance": { "type": "number" },
          "elevationGain": { "type": "number" },
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Racer": {
        "type": "object",
        "required": ["id", "firstName", "lastName", "profileImageUrl", "category", "ageGroup", "sexCategory", "registeredAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "profileImageUrl": { "type": "string" },
          "category": { "type": "string", "enum": ["MTB", "EBIKE", "BOTH"] },
          "ageGroup": { "type": "string", "enum": ["U30", "30-39", "40-49", "50-59", "60+"] },
          "sexCategory": { "type": "string", "enum": ["MALE", "FEMALE"] },
          "registeredAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateChallengeRequest": {
        "type": "object",
        "required": ["name", "description", "startDate", "endDate"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" }
        }
      },
      "AddSegmentRequest": {
        "type": "object",
        "required": ["challengeId", "stravaSegmentId", "name", "distance", "elevationGain"],
        "properties": {
          "challengeId": { "type": "string", "format": "uuid" },
          "stravaSegmentId": { "type": "integer" },
          "name": { "type": "string" },
          "distance": { "type": "number" },
          "elevationGain": { "type": "number" },
          "description": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "country": { "type": "string" },
          "averageGrade": { "type": "number" },
          "maximumGrade": { "type": "number" },
          "elevationHigh": { "type": "number" },
          "elevationLow": { "type": "number" },
          "climbCategory": { "type": "integer" },
          "private": { "type": "boolean" },
          "hazardous": { "type": "boolean" },
          "starCount": { "type": "integer" },
          "athleteCount": { "type": "integer" },
          "effortCount": { "type": "integer" },
          "rawStravaMetadata": { "type": "object" }
        }
      },
      "UpdateRacerRequest": {
        "type": "object",
        "properties": {
          "category": { "type": "string", "enum": ["MTB", "EBIKE", "BOTH"] },
          "ageGroup": { "type": "string", "enum": ["U30", "30-39", "40-49", "50-59", "60+"] },
          "sexCategory": { "type": "string", "enum": ["MALE", "FEMALE"] }
        },
        "minProperties": 1
      },
...
      "ConnectedAthlete": {
        "type": "object",
        "required": ["racerId", "stravaAthleteId", "name", "profileImageUrl", "category", "ageGroup", "sexCategory"],
        "properties": {
          "racerId": { "type": "string", "format": "uuid" },
          "stravaAthleteId": { "type": "integer" },
          "name": { "type": "string" },
          "profileImageUrl": { "type": "string" },
          "category": { "type": "string", "enum": ["MTB", "EBIKE", "BOTH"] },
          "ageGroup": { "type": "string", "enum": ["U30", "30-39", "40-49", "50-59", "60+"] },
          "sexCategory": { "type": "string", "enum": ["MALE", "FEMALE"] }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid admin JWT.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/challenges": {
      "get": {
        "tags": ["Admin"],
        "summary": "List challenges",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "All challenges.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["challenges"],
                  "properties": {
                    "challenges": { "type": "array", "items": { "$ref": "#/components/schemas/Challenge" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create a challenge",
        "security": [{ "AdminBearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChallengeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/admin/segments": {
      "get": {
        "tags": ["Admin"],
        "summary": "List active challenge segments",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Admin segment list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["segments"],
                  "properties": {
                    "segments": { "type": "array", "items": { "$ref": "#/components/schemas/Segment" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Add a segment to a challenge",
        "security": [{ "AdminBearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddSegmentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Segment created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {
                    "id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/segments/{segmentId}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a segment",
        "security": [{ "AdminBearerAuth": [] }],
        "parameters": [
          { "name": "segmentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Segment deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["deleted"],
                  "properties": {
                    "deleted": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/racers": {
      "get": {
        "tags": ["Admin"],
        "summary": "List active challenge racers",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Admin racer list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["racers"],
                  "properties": {
                    "racers": { "type": "array", "items": { "$ref": "#/components/schemas/Racer" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/racers/{racerId}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update a racer's bike, sex, or age category",
        "security": [{ "AdminBearerAuth": [] }],
        "parameters": [
          { "name": "racerId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRacerRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated racer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["racer"],
                  "properties": {
                    "racer": { "$ref": "#/components/schemas/Racer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid category value.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/connected-athletes": {
      "get": {
        "tags": ["Admin"],
        "summary": "List racers with stored Strava tokens",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Connected athletes consuming Strava athlete slots.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["athletes"],
                  "properties": {
                    "athletes": { "type": "array", "items": { "$ref": "#/components/schemas/ConnectedAthlete" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/connected-athletes/cleanup": {
      "post": {
        "tags": ["Admin"],
        "summary": "Revoke and delete all stored Strava connections",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Cleanup results.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["results"],
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["racerId", "status"],
                        "properties": {
                          "racerId": { "type": "string" },
                          "stravaAthleteId": { "type": "integer" },
                          "status": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/admin/challenges/{challengeId}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a challenge with cascade (segments, results, leaderboard entries)",
        "security": [{ "AdminBearerAuth": [] }],
        "parameters": [
          { "name": "challengeId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Challenge and related data deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["deleted", "segmentsDeleted"],
                  "properties": {
                    "deleted": { "type": "boolean" },
                    "segmentsDeleted": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/segments/all": {
      "get": {
        "tags": ["Admin"],
        "summary": "List segments across all challenges",
        "security": [{ "AdminBearerAuth": [] }],
        "responses": {
          "200": {
            "description": "All segments with challenge name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["segments"],
                  "properties": {
                    "segments": { "type": "array", "items": { "$ref": "#/components/schemas/Segment" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}